Hl/support v100 mlu290

See merge request jiangyongjiu1/yolov3!10

+++ utils/plots.py
@@ -20,6 +20,7 @@ from PIL import Image, ImageDraw, ImageFont
 from utils.general import (LOGGER, Timeout, check_requirements, clip_coords, increment_path, is_ascii, is_chinese,
                            try_except, user_config_dir, xywh2xyxy, xyxy2xywh)
 from utils.metrics import fitness
+from utils.config import int_dtype
 
 # Settings
 CONFIG_DIR = user_config_dir()  # Ultralytics settings dir
@@ -460,7 +461,7 @@ def save_one_box(xyxy, im, file='image.jpg', gain=1.02, pad=10, square=False, BG
     if square:
         b[:, 2:] = b[:, 2:].max(1)[0].unsqueeze(1)  # attempt rectangle to square
     b[:, 2:] = b[:, 2:] * gain + pad  # box wh * gain + pad
-    xyxy = xywh2xyxy(b).long()
+    xyxy = xywh2xyxy(b).to(dtype=int_dtype)
     clip_coords(xyxy, im.shape)
     crop = im[int(xyxy[0, 1]):int(xyxy[0, 3]), int(xyxy[0, 0]):int(xyxy[0, 2]), ::(1 if BGR else -1)]
     if save:
